perf(live): spend fewer tokens per turn and less work per frame
Only the leading components carry a worked example, the per-turn prompt
sits inside the cached prefix on Anthropic, a repair pass no longer resends
the component menu, drawer blocks are generated when the drawer is opened,
and a prefetched suggestion is reused across turns instead of being thrown
away and paid for twice. A glimpse is billed as a glimpse rather than at a
reasoning model's floor.

Spotlight and pen-gesture teaching stays on lean turns: withholding it
saved ~1.2k tokens and quietly removed marks from a large share of answers.

Ambient loops now pause on a hidden tab and off-screen, the on-device
semantic model is fetched on composer intent rather than any keypress, and
local stores write through one shared quota ledger so the store that writes
last no longer loses data silently.

+++ tests/live-scrub-unbuild.test.ts
@@ -8,7 +8,7 @@ import { sayable } from '../src/voice/tts';
 // track's end the canvas is whole; reveal-time marks remain the floor for slow streams.
 
 const audio: TurnAudio = {
-  pcm: new Float32Array(0),
+  pcm: new Int16Array(0),
   sampleRate: 24000,
   duration: 20,
   spans: [
